remove from root

+++ week02/src/main/java/bbk/Sensor.java
@@ -2,5 +2,6 @@ package bbk;
 
 public interface Sensor extends SecuritySensor {
 
-  double getBatteryPercentage();
+  int getBatteryPercentage();
+
 }
